requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.2.3
Flask-Migrate

